pyproject.toml
[tool.poetry]
name = "dating_site"
version = "0.1.0"
description = ""
authors = ["Sergey_Sharay <sergeysharay1987@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0.2"
Pillow = "^9.0.1"
djangorestframework = "^3.13.1"
django-filter = "^22.1"
psycopg2 = "^2.9.4"
dj-rest-auth = {extras = ["with_social"], version = "^2.2.5"}
django-username-email = "^2.5.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
pytest-django = "^4.5.2"
model-bakery = "^1.9.0"
pre-commit = "^2.20.0"

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
